third_party/WebKit/LayoutTests/http/tests/media/media-source/webkitmediasource-removesourcebuffer-texttrack.html

<!DOCTYPE html>
<html>
    <head>
        <script src="/media-resources/video-test.js"></script>
        <script>
            if (window.testRunner)
                testRunner.dumpAsText();

            var mediaSource = new WebKitMediaSource();
            var sourceBuffer;
            var textTrack;

            mediaSource.addEventListener('webkitsourceopen', function()
            {
                var xhr = new XMLHttpRequest();

                xhr.open('GET', '../resources/media-source/webm/test-vp8-vorbis-webvtt.webm', true);
                xhr.responseType = 'arraybuffer';
                xhr.send();

                xhr.onload = function()
                {
                    if (xhr.status != 200 && xhr.status != 0) {
                        failTest("HTTP request failed");
                        return;
                    }

                    sourceBuffer = mediaSource.addSourceBuffer('video/webm; codecs="vorbis,vp8"');
                    sourceBuffer.append(new Uint8Array(xhr.response));
                    mediaSource.endOfStream();
                };
            }, false);

            function onLoad()
            {
                var v = document.getElementById("vid");

                waitForEvent('loadedmetadata', function()
                {
                    textTrack = v.textTracks[0];
                    mediaSource.removeSourceBuffer(sourceBuffer);
                    v.src = "";
                }, false, false, v, false);

                waitForEvent('error', function()
                {
                    if (v.error.code == v.error.MEDIA_ERR_SRC_NOT_SUPPORTED) {
                        textTrack.mode = "disabled";
                        endTest();
                    } else {
                        failTest("Expected error code MEDIA_ERR_SRC_NOT_SUPPORTED");
                    }
                }, false, false, v, false);

                v.src = window.URL.createObjectURL(mediaSource);
            }
        </script>
    </head>
    <body onload="onLoad()">
        <video id="vid"> </video>
    </body>
</html>
